requirements.txt
matplotlib==3.6.3
numpy==2.2.6
pygame==2.6.1
PyYAML==6.0.2
scipy==1.15.3
